Add vagrant support for running tox
The fact that tox covers Python2.4 --> Python 3.2 is a pain. It relies on a specific setup
allowing installation of ZeroMQ, SQLAlchemy and more, and also a vanilla installation of Python 2.4 and 2.5.

The Vagrant image takes a while to load, but at least gives a clean, agreed-upon environment in which tests must succeed.
It establishes an Ubuntu 12.04, 64-bit image, with deadsnakes Pythons, and runs tox inside it.

It can be easily invoked with "make vagrant_toxtest"