SyslogHandler: Reconnect to TCP sockets when they are disconnected
When SyslogHandler is disconnected from a TCP socket, e.g., because of
networking issues or because the syslog server at the other end of the
socket was restarted, attempt to reconnect.

Note that this doesn't mean that no messages will be lost. Obviously,
if the remote server is down for a while, then messages will be lost
while we're getting "Connection Refused" trying to reconnect to it.
Even if the disconnection is just a temporary blip, a few messages may
be lost because the kernel doesn't always notice right away when a TCP
connection goes away.

While fixing this issue, I noticed another oddity which I fixed. After
SyslogHandler connects to a network socket, it fetches the local
address of the connection of the connection and assigns it to
self.address, whose value was previously the _remote_ address that we
were trying to connect to. Overwriting the remote address made it
difficult (i.e., impossible!) to reconnect when the connection was
interrupted. I didn't want to just remove the behavior of overwriting
self.address, because there may be consumers of Logbook who depend on
it. Therefore, I changed the code to save the remote address in
self.remote_address and use that, rather than self.address, when
connecting.