SyslogHandler: Handle multiline messages and exceptions properly
My previous attempts to handle delimiters properly for AF_INET logging
and use RFC 5424 format for such messages had two problems:

1. StringFormatter appends exceptions to messages _after_ formatting
them, as opposed to including the exception in record.message before
formatting. As a result, with my changes in place, when exceptions
were being logged, the log message as transmitted to the syslog server
was not ending with a record delimiter as it should have been.

2. Messages with embedded record delimiters need to have each segment
of the message prefixed with the correct RFC 5424 header.

Therefore, the approach of having the StringFormatter do the RFC 5424
header formatting is herein abandoned. Instead, we format the message,
then split it on the record delimiter, then put the RFC 5424 at the
beginning and the record delimiter at the end of each segment of the
message.

Each of these segments is then sent separately to the socket, rather
than concatenating all of them, which is unnecessary overhead.