Use the channel as the application name when it makes sense
Before we were using RFC 5424 format for AF_INET logging, the log
channel was serving as the application name for all intents and
purposes, when the application name wasn't being specified. However,
in the world of RFC 5424, that's no longer the case, since the
application name is a separate field in the RFC header.

Therefore, if the user didn't specify an application name when
creating the SyslogHandler, the channel name should be treated as the
application name when emitting AF_INET logging.