The big context change
The old contextvars support just didn't work, so I've been working on
this rewrite for a while. A working contextvars implementation is able
to replace push_thread since they are thread-local, and is able to
replace push_greenlet because greenlet supports contextvars.

The previous gevent interfaces are entirely deprecated by this patch,
with gevent.monkey being the supported way forward. That is to say,
I don't think the couple of locks we have being gevent locks is
sufficient to be fully gevent compatible, since we don't control a lot
of the code we interface with.